+++ test/cli/test_commands.py
@@ -297,6 +297,27 @@ def test_unsatisfiable(cli, schema_url):
     assert "hypothesis.errors.Unsatisfiable: Unable to satisfy schema parameters for this endpoint" in lines
 
 
+@pytest.mark.endpoints("flaky")
+def test_flaky(cli, schema_url):
+    # When the endpoint fails / succeeds randomly
+    result = cli.run_inprocess(schema_url)
+    # Then the whole Schemathesis run should fail
+    assert result.exit_code == 1
+    # And standard Hypothesis error should not appear in the output
+    assert "Failed to reproduce exception. Expected:" not in result.stdout
+    # And this endpoint should be marked as errored in the progress line
+    assert "GET /api/flaky E" in result.stdout
+    # And it should be displayed only once in "ERRORS" section
+    assert "= ERRORS =" in result.stdout
+    assert "_ GET: /api/flaky _" in result.stdout
+    # And it should not go into "FAILURES" section
+    assert "= FAILURES =" not in result.stdout
+    # And more clear error message is displayed instead of Hypothesis one
+    lines = result.stdout.split("\n")
+    assert "hypothesis.errors.Flaky: Tests on this endpoint produce unreliable results: " in lines
+    assert "Falsified on the first call but did not on a subsequent one" in lines
+
+
 @pytest.mark.endpoints("invalid")
 def test_invalid_endpoint(cli, schema_url):
     # When the app's schema contains errors